backend-python-django/requirements.txt
djangorestframework==3.12.4
PyYAML==5.4.1